// src/interfaces/SearchPath.ts
export class SearchPath {
    nodeOne: string | undefined;
    nodeTwo: string | undefined;
    isComponentOf: boolean = false;
    isSemanticComponentOf: boolean = false;
    isPictographicRadicalOf: boolean = false;
    isPhoneticRadicalOf: boolean = false;
    isInitialOf: boolean = false;
    isFinalOf: boolean = false;
    isToneOf: boolean = false;
    isTransformedTo: boolean = false;

    constructor() { }
}